requirements.txt
Django>=4.2
djangorestframework>=3.14

psycopg2-binary>=2.9
pgvector>=0.2.4

celery>=5.3
redis>=5.0

langchain>=0.1
langgraph>=0.0.30
langchain-groq>=0.0.3

groq>=0.5.0

sentence-transformers>=2.7
torch>=2.1

PyMuPDF>=1.23
python-pptx>=0.6.23

reportlab>=4.0

gunicorn>=21.2

python-dotenv>=1.0
